% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bddensity.R
\name{bddensity}
\alias{bddensity}
\title{estimate density curve}
\usage{
bddensity(
  d,
  list_beta = c(0, 0.5, 1, 1.5, 2),
  beta0 = 1,
  m = 400,
  log.range = c(-4, 1),
  xlim = c(NA, NA),
  data.is.integer = F
)
}
\arguments{
\item{d}{data}

\item{list_beta}{beta}

\item{beta0}{beta}

\item{m}{beta}

\item{log.range}{beta}

\item{xlim}{xilm}
}
\description{
estimate density curve
}
