% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdquantile.R
\name{bdquantile}
\alias{bdquantile}
\title{plot density curve}
\usage{
bdquantile(result_bddesntiy, q = 0.05, n_BT = 1000, xlim = c(NA, NA), m = 400)
}
\arguments{
\item{q}{beta}

\item{n_BT}{beta}

\item{xlim}{beta}

\item{m}{xilm}

\item{result_bddensity}{result_bddensity}
}
\description{
plot density curve
}
