% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diagnosis}
\alias{diagnosis}
\title{diagnosis}
\usage{
diagnosis(result_bddensity, result_bdquantile, alpha = 0.95, persistence = 5)
}
\arguments{
\item{result_bddensity}{result_bddensity}

\item{result_bdquantile}{result_bdquantile}

\item{alpha}{alpha}

\item{persistence}{persistence}
}
\description{
diagnosis
}
