% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_peaks.R
\name{fit_peaks}
\alias{fit_peaks}
\title{fit_peaks}
\usage{
fit_peaks(result_bddenstiy, result_bdquantile)
}
\arguments{
\item{result_bddenstiy}{result_bddenstiy}

\item{result_bdquantile}{result_bdquantile}
}
\description{
fit_peaks
}
