% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_fit.R
\name{peak_fit}
\alias{peak_fit}
\title{peak fitting}
\usage{
peak_fit(xy, result_bdquantile = NULL, alpha = 0.95)
}
\arguments{
\item{xy}{xy}

\item{result_bdquantile}{result_bdquantile}
}
\description{
peak fitting
}
