% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(
  result_bddensity,
  result_bdquantile = NULL,
  xlim = c(NA, NA),
  xlab = "x",
  ylab = "frequency",
  color = rgb(250/255, 128/255, 114/255, alpha = 0.3),
  rug = T
)
}
\arguments{
\item{result_bddensity}{result_bddensity,}

\item{result_bdquantile}{result_bdquantile}

\item{xlim}{xlim}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{color}{color}

\item{rug}{rug}

\item{bdq}{bdq}
}
\description{
plot_density
}
