% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_peaks.R
\name{plot_fit_peaks}
\alias{plot_fit_peaks}
\title{plot_fit_peaks}
\usage{
plot_fit_peaks(result_peak_fit, alpha = 0.95)
}
\arguments{
\item{result_peak_fit}{result_peak_fit}

\item{alpha}{alpha}
}
\description{
plot_fit_peaks
}
