% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak_fit.R
\name{plot_peak_fit}
\alias{plot_peak_fit}
\title{plot_peak_fit}
\usage{
plot_peak_fit(result_peak_fit, alpha = 0.95)
}
\arguments{
\item{result_peak_fit}{result_peak_fit}

\item{alpha}{alpha}
}
\description{
plot_peak_fit
}
