% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments_peaks.R
\name{segments_peaks}
\alias{segments_peaks}
\title{segments_peaks}
\usage{
segments_peaks(result_find_peaks, color = "royalblue", lwd = 2)
}
\arguments{
\item{result_find_peaks}{result_find_peaks}

\item{color}{color}

\item{lwd}{lwd}
}
\description{
segments_peaks
}
