% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdplot.R
\name{bdplot}
\alias{bdplot}
\title{plot density curve}
\usage{
bdplot(
  d,
  xlim = c(NA, NA),
  m = 500,
  CI = T,
  method = "auto",
  xlab = "x",
  ylab = "frequency",
  beta0 = 1,
  list_beta = c(0, 0.5, 1, 1.5, 2),
  color = rgb(250/255, 128/255, 114/255, alpha = 0.3),
  rug = T
)
}
\arguments{
\item{d}{data}

\item{xlim}{beta}

\item{m}{beta}

\item{CI}{beta}

\item{method}{beta}

\item{xlab}{xilm}

\item{ylab}{xilm}

\item{beta0}{xilm}

\item{list_beta}{xilm}

\item{color}{xilm}

\item{rug}{xilm}
}
\description{
plot density curve
}
