% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdquantile.R
\name{bdquantile}
\alias{bdquantile}
\title{plot density curve}
\usage{
bdquantile(d, h, q = 0.05, n_BT = 1000, xlim = c(NA, NA), m = 400)
}
\arguments{
\item{d}{data}

\item{h}{beta}

\item{q}{beta}

\item{n_BT}{beta}

\item{xlim}{beta}

\item{m}{xilm}
}
\description{
plot density curve
}
